Progress on issue 486: New setting to replace "Auto Drum SP" --> "Drum SP Mode":
 * "Auto / Fills" (default) - Will use drum fills for starpower activation if markings
exist, otherwise it will be activated automatically when there is enough.
 * "Manual / Fills" - Will use drum fills for starpower activation if markings exist,
otherwise it will require a button press to activate.

Progress on issue 102:
Last bit of missing logic was added to prevent the final case in which gaining starpower while you already had over 50% would cause an incoming drum fill to vanish.

git-svn-id: http://fofix.googlecode.com/svn/MFH-Mod/trunk@857 48f48816-a574-47fe-9a59-6914b3aef992