Progress on issue 560 - ...this is working great for me in a Windows XP 32-bit source environment.  I'm also now seeing around 150 FPS in menus instead of the 60 I'm used to on my laptop running at minimum speed.
I've added a quick and dirty hack to fix all the empty font boxes from showing up when the special glyph characters are used.  Now, all Left and Right arrow icons will show up as "<" and ">" - this works great.  Stars will show up in the songlist as asterisks *, and do not quite line up right, but they work fine.  This will allow the alpha testers to experiment with this, and we will fix the ugly asterisk-*-stars in a later revision as long as this doesn't break the game for anyone.



git-svn-id: http://fofix.googlecode.com/svn/MFH-Mod/trunk@1002 48f48816-a574-47fe-9a59-6914b3aef992