Remove the main loop exception handler.
We should never have to deal with an exception from game code.  If game
code raises an exception, the game should crash, and then we can find
the offending code and *fix it*.  How do we know it's safe to just report
the exception, eat it, and continue in all cases?