Implemented Solo Sidebars
Had to add solo overdrive sidebars, because when the sidebar images are rendered they are replaced not added on top of each other. So this second image will be used for having both solo sidebars and od sidebars at once. The solo specific odsb's will only be rendered if the solo sidebars are present.